Fixed Dapapath.map_vol  function
Dapapath.map_vol function does the mapping only if chained=False but
skip it in any other cases.  Mapping is still required for creation of
an image as there is a bug with qemu-img. It can't create QCow2
formatted image over sheepdog vdi if we specify the image as uri like
`qemu-img create -f qcow2 sheepdog:image-name 1G` but it works with
previously created and mapped as block device sheepdog vdi